Fixed test & added form data to extracted data
First off, now, you have a button to update the form content, and
it is also shown now on the page itself.

The extracted data now also contains some data about the form,
such as the location of the form, the method it'll use, and the
page that it will GET/POST to.